+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 #### 功能构建
 
-- (无)
+- 新建 Controller。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.finance.BillFileController。
 
 #### Bug修复